+++ amico/__init__.py
@@ -8,5 +8,8 @@ from . import lut
 from . import models
 from . import util
 
-from pkg_resources import get_distribution
-__version__ = get_distribution('dmri-amico').version
+try:
+    from importlib.metadata import version
+except ImportError:
+    from importlib_metadata import version
+__version__ = version('dmri-amico')