+++ amico/core.py
@@ -4,8 +4,8 @@ import numpy as np
 import time
 import glob
 import sys
-from os import makedirs, remove
-from os.path import exists, join as pjoin, isfile, isdir
+from os import makedirs, remove, cpu_count
+from os.path import exists, join as pjoin, isfile
 import inspect
 
 import nibabel
@@ -17,13 +17,12 @@ import amico.models
 from amico.lut import is_valid, valid_dirs
 from dipy.core.gradients import gradient_table
 import dipy.reconst.dti as dti
-from amico.util import PRINT, LOG, NOTE, WARNING, ERROR, get_verbose
+from amico.util import PRINT, LOG, WARNING, ERROR, get_verbose
+from dicelib.ui import ProgressBar
 from pkg_resources import get_distribution
-from joblib import Parallel, delayed, cpu_count, wrap_non_picklable_objects
-from tqdm import tqdm
+from threadpoolctl import ThreadpoolController
 
-
-def setup( lmax=12, ndirs=None ) :
+def setup( lmax=12 ) :
     """General setup/initialization of the AMICO framework.
 
     Parameters
@@ -31,14 +30,13 @@ def setup( lmax=12, ndirs=None ) :
     lmax : int
         Maximum SH order to use for the rotation phase (default : 12).
         NB: change only if you know what you are doing.
-     ndirs : int
-        DEPRECATED. Now, all directions are precomputed.
     """
-    if ndirs is not None:
-        WARNING('"ndirs" parameter is deprecated')
     LOG( '\n-> Precomputing rotation matrices:' )
-    for n in tqdm(valid_dirs(), ncols=70, bar_format='   |{bar}| {percentage:4.1f}%', disable=(get_verbose()<3)):
-        amico.lut.precompute_rotation_matrices( lmax, n )
+    dirs = valid_dirs()
+    with ProgressBar(total=len(dirs), disable=get_verbose()<3) as pbar:
+        for dir in dirs:
+            amico.lut.precompute_rotation_matrices(lmax, dir)
+            pbar.update()
     LOG('   [ DONE ]')
 
 
@@ -48,7 +46,7 @@ class Evaluation :
     evaluation with the AMICO framework.
     """
 
-    def __init__( self, study_path, subject, output_path=None ) :
+    def __init__( self, study_path='.', subject='.', output_path=None ) :
         """Setup the data structure with default values.
 
         Parameters
@@ -61,16 +59,20 @@ class Evaluation :
             Optionally sets a custom full path for the output. Leave as None
             for default behaviour - output in study_path/subject/AMICO/<MODEL>
         """
-        self.niiDWI      = None # set by "load_data" method
-        self.niiDWI_img  = None
-        self.scheme      = None
-        self.niiMASK     = None
-        self.niiMASK_img = None
-        self.model       = None # set by "set_model" method
-        self.KERNELS     = None # set by "load_kernels" method
-        self.RESULTS     = None # set by "fit" method
-        self.mean_b0s    = None # set by "load_data" method
-        self.htable      = None
+        self.niiDWI        = None    # set by "load_data" method
+        self.niiDWI_img    = None
+        self.scheme        = None
+        self.niiMASK       = None
+        self.niiMASK_img   = None
+        self.model         = None    # set by "set_model" method
+        self.KERNELS       = None    # set by "load_kernels" method
+        self.y             = None    # set by "fit" method
+        self.DIRs          = None    # set by "fit" method
+        self.nthreads      = None    # set by "fit" method
+        self.BLAS_nthreads = None    # set by "generate_kernel", "load_kernels" and "fit" methods
+        self.RESULTS       = None    # set by "fit" method
+        self.mean_b0s      = None    # set by "load_data" method
+        self.htable        = None
 
         # store all the parameters of an evaluation with AMICO
         self.CONFIG = {}
@@ -82,14 +84,20 @@ class Evaluation :
 
         self.set_config('peaks_filename', None)
         self.set_config('doNormalizeSignal', True)
-        self.set_config('doKeepb0Intact', False)        # does change b0 images in the predicted signal
+        self.set_config('doKeepb0Intact', False)       # does change b0 images in the predicted signal
+        self.set_config('doComputeRMSE', False)
         self.set_config('doComputeNRMSE', False)
-        self.set_config('doSaveCorrectedDWI', False)
-        self.set_config('doMergeB0', False)             # Merge b0 volumes
-        self.set_config('doDebiasSignal', False)        # Flag to remove Rician bias
-        self.set_config('DWI-SNR', None)                # SNR of DWI image: SNR = b0/sigma
-        self.set_config('doDirectionalAverage', False)  # To perform the directional average on the signal of each shell
-        self.set_config('parallel_jobs', -1)            # Number of jobs to be used in multithread-enabled parts of code
+        self.set_config('doSaveModulatedMaps', False)  # NODDI model specific config
+        self.set_config('doSaveCorrectedDWI', False)   # FreeWater model specific config
+        self.set_config('doMergeB0', False)            # Merge b0 volumes
+        self.set_config('doDebiasSignal', False)       # Flag to remove Rician bias
+        self.set_config('DWI-SNR', None)               # SNR of DWI image: SNR = b0/sigma
+        self.set_config('doDirectionalAverage', False) # To perform the directional average on the signal of each shell
+        self.set_config('nthreads', -1)                # Number of threads to be used in multithread-enabled parts of code (default: -1)
+        self.set_config('DTI_fit_method', 'OLS')       # Fit method for the Diffusion Tensor model (dipy) (default: 'OLS')
+        self.set_config('BLAS_nthreads', 1)            # Number of threads used in the threadpool-backend of common BLAS implementations (dafault: 1)
+
+        self._controller = ThreadpoolController()
 
     def set_config( self, key, value ) :
         self.CONFIG[ key ] = value
@@ -116,7 +124,6 @@ class Evaluation :
         replace_bad_voxels : float, optional
             Value to be used to fill NaN and Inf values in the signal. (default : do nothing)
         """
-
         # Loading data, acquisition scheme and mask (optional)
         LOG( '\n-> Loading data:' )
         tic = time.time()
@@ -127,7 +134,7 @@ class Evaluation :
         self.set_config('dwi_filename', dwi_filename)
         self.set_config('b0_min_signal', b0_min_signal)
         self.set_config('replace_bad_voxels', replace_bad_voxels)
-        self.niiDWI  = nibabel.load( pjoin(self.get_config('DATA_path'), dwi_filename) )
+        self.niiDWI = nibabel.load( pjoin(self.get_config('DATA_path'), dwi_filename) )
         self.niiDWI_img = self.niiDWI.get_fdata().astype(np.float32)
         hdr = self.niiDWI.header if nibabel.__version__ >= '2.0.0' else self.niiDWI.get_header()
         if self.niiDWI_img.ndim != 4 :
@@ -171,7 +178,7 @@ class Evaluation :
         if mask_filename is not None :
             if not isfile( pjoin(self.get_config('DATA_path'), mask_filename) ):
                 ERROR( 'MASK file not found' )
-            self.niiMASK  = nibabel.load( pjoin( self.get_config('DATA_path'), mask_filename) )
+            self.niiMASK = nibabel.load( pjoin( self.get_config('DATA_path'), mask_filename) )
             self.niiMASK_img = self.niiMASK.get_fdata().astype(np.uint8)
             niiMASK_hdr = self.niiMASK.header if nibabel.__version__ >= '2.0.0' else self.niiMASK.get_header()
             PRINT('\t\t- dim    = %d x %d x %d' % self.niiMASK_img.shape[:3])
@@ -294,7 +301,7 @@ class Evaluation :
 
 
     def set_solver( self, **params ) :
-        """Setup the specific parameters of the solver to fit the model.
+        """Set up the specific parameters of the solver to fit the model.
         Dispatch to the proper function, depending on the model; a model should provide a "set_solver" function to set these parameters.
         Currently supported parameters are:
         StickZeppelinBall:      'set_solver()' not implemented
@@ -307,7 +314,7 @@ class Evaluation :
         """
         if self.model is None :
             ERROR( 'Model not set; call "set_model()" method first' )
-
+        
         solver_params = list(inspect.signature(self.model.set_solver).parameters)
         params_new = {}
         for key in params.keys():
@@ -316,10 +323,11 @@ class Evaluation :
             else:
                 params_new[key] = params[key]
 
-        self.set_config('solver_params', self.model.set_solver( **params_new ))
+        self.model.set_solver(**params_new)
+        self.set_config('solver_params', params_new)
 
 
-    def generate_kernels( self, regenerate = False, lmax = 12, ndirs = 32761 ) :
+    def generate_kernels( self, regenerate = False, lmax = 12, ndirs = 500 ) :
         """Generate the high-resolution response functions for each compartment.
         Dispatch to the proper function, depending on the model.
 
@@ -330,14 +338,16 @@ class Evaluation :
         lmax : int
             Maximum SH order to use for the rotation procedure (default : 12)
         ndirs : int
-            Number of directions on the half of the sphere representing the possible orientations of the response functions (default : 32761)
-         """
+            Number of directions on the half of the sphere representing the possible orientations of the response functions (default : 500)
+        """
         if self.scheme is None :
             ERROR( 'Scheme not loaded; call "load_data()" first' )
         if self.model is None :
             ERROR( 'Model not set; call "set_model()" method first' )
         if not is_valid(ndirs):
-            ERROR( 'Unsupported value for ndirs.\nNote: Supported values for ndirs are [1, 500, 1000, 1500, 2000, 2500, 3000, 3500, 4000, 4500, 5000, 5500, 6000, 6500, 7000, 7500, 8000, 8500, 9000, 9500, 10000, 32761 (default)]' )
+            ERROR( 'Unsupported value for ndirs.\nNote: Supported values for ndirs are [1, 500 (default), 1000, 1500, 2000, 2500, 3000, 3500, 4000, 4500, 5000, 5500, 6000, 6500, 7000, 7500, 8000, 8500, 9000, 9500, 10000, 32761]' )
+        
+        self.BLAS_nthreads = self.get_config('BLAS_nthreads') if self.get_config('BLAS_nthreads') > 0 else cpu_count() if self.get_config('BLAS_nthreads') == -1 else ERROR('Number of BLAS threads must be positive or -1')
 
         # store some values for later use
         self.set_config('lmax', lmax)
@@ -364,7 +374,8 @@ class Evaluation :
 
         # Dispatch to the right handler for each model
         tic = time.time()
-        self.model.generate( self.get_config('ATOMS_path'), aux, idx_IN, idx_OUT, ndirs )
+        with self._controller.limit(limits=self.BLAS_nthreads, user_api='blas'):
+            self.model.generate( self.get_config('ATOMS_path'), aux, idx_IN, idx_OUT, ndirs )
         LOG( f'   [ {time.time() - tic:.1f} seconds ]' )
 
 
@@ -376,6 +387,8 @@ class Evaluation :
             ERROR( 'Model not set; call "set_model()" method first' )
         if self.scheme is None :
             ERROR( 'Scheme not loaded; call "load_data()" first' )
+        
+        self.BLAS_nthreads = self.get_config('BLAS_nthreads') if self.get_config('BLAS_nthreads') > 0 else cpu_count() if self.get_config('BLAS_nthreads') == -1 else ERROR('Number of BLAS threads must be positive or -1')
 
         tic = time.time()
         LOG( f'\n-> Resampling LUT for subject "{self.get_config("subject")}":' )
@@ -387,66 +400,16 @@ class Evaluation :
         self.htable = amico.lut.load_precomputed_hash_table( self.get_config('ndirs') )
 
         # Dispatch to the right handler for each model
-        self.KERNELS = self.model.resample( self.get_config('ATOMS_path'), idx_OUT, Ylm_OUT, self.get_config('doMergeB0'), self.get_config('ndirs') )
+        with self._controller.limit(limits=self.BLAS_nthreads, user_api='blas'):
+            self.KERNELS = self.model.resample( self.get_config('ATOMS_path'), idx_OUT, Ylm_OUT, self.get_config('doMergeB0'), self.get_config('ndirs') )
 
         LOG( f'   [ {time.time() - tic:.1f} seconds ]')
 
 
     def fit( self ) :
-        """Fit the model to the data iterating over all voxels (in the mask) one after the other.
+        """Fit the model to the data.
         Call the appropriate fit() method of the actual model used.
         """
-
-        @delayed
-        @wrap_non_picklable_objects
-        def fit_voxel(self, ix, iy, iz, dirs, DTI) :
-            """Perform the fit in a single voxel.
-            """
-            # prepare the signal
-            y = self.niiDWI_img[ix, iy, iz, :].astype(np.float64)
-            y[y < 0] = 0  # [NOTE] this should not happen!
-
-            # fitting directions if not
-            if not self.get_config('doDirectionalAverage') and DTI is not None :
-                dirs = DTI.fit( y ).directions[0].reshape(-1, 3)
-
-            # dispatch to the right handler for each model
-            results, dirs, x, A = self.model.fit( y, dirs, self.KERNELS, self.get_config('solver_params'), self.htable)
-
-            # compute fitting error
-            if self.get_config('doComputeNRMSE') :
-                y_est = np.dot( A, x )
-                den = np.sum(y**2)
-                NRMSE = np.sqrt( np.sum((y-y_est)**2) / den ) if den > 1e-16 else 0
-            else :
-                NRMSE = 0.0
-
-            y_fw_corrected = None
-            if self.get_config('doSaveCorrectedDWI') :
-
-                if self.model.name == 'Free-Water' :
-                    n_iso = len(self.model.d_isos)
-
-                    # keep only FW components of the estimate
-                    x[0:x.shape[0]-n_iso] = 0
-
-                    # y_fw_corrected below is the predicted signal by the anisotropic part (no iso part)
-                    y_fw_part = np.dot( A, x )
-
-                    # y is the original signal
-                    y_fw_corrected = y - y_fw_part
-                    y_fw_corrected[ y_fw_corrected < 0 ] = 0 # [NOTE] this should not happen!
-
-                    if self.get_config('doNormalizeSignal') and self.scheme.b0_count > 0 :
-                        y_fw_corrected = y_fw_corrected * self.mean_b0s[ix,iy,iz]
-
-                    if self.get_config('doKeepb0Intact') and self.scheme.b0_count > 0 :
-                        # put original b0 data back in.
-                        y_fw_corrected[self.scheme.b0_idx] = y[self.scheme.b0_idx]*self.mean_b0s[ix,iy,iz]
-
-            return results, dirs, NRMSE, y_fw_corrected
-
-
         if self.niiDWI is None :
             ERROR( 'Data not loaded; call "load_data()" first' )
         if self.model is None :
@@ -455,77 +418,86 @@ class Evaluation :
             ERROR( 'Response functions not generated; call "generate_kernels()" and "load_kernels()" first' )
         if self.KERNELS['model'] != self.model.id :
             ERROR( 'Response functions were not created with the same model' )
-        n_jobs = self.get_config( 'parallel_jobs' )
-        if n_jobs == -1 :
-            n_jobs = cpu_count()
-        elif n_jobs == 0 or n_jobs < -1:
-            ERROR( 'Number of parallel jobs must be positive or -1' )
+        if self.get_config('DTI_fit_method') not in ['OLS', 'LS', 'WLS', 'NLLS', 'RT', 'RESTORE', 'restore']:
+            ERROR("DTI fit method must be one of the following:\n'OLS'(default) or 'LS': ordinary least squares\n'WLS': weighted least squares\n'NLLS': non-linear least squares\n'RT' or 'RESTORE' or 'restore': robust tensor\nNOTE: more info at https://dipy.org/documentation/1.6.0./reference/dipy.reconst/#dipy.reconst.dti.TensorModel")
+        
+        self.nthreads = self.get_config('nthreads') if self.get_config('nthreads') > 0 else cpu_count() if self.get_config('nthreads') == -1 else ERROR('Number of parallel threads must be positive or -1')
+        self.BLAS_nthreads = self.get_config('BLAS_nthreads') if self.get_config('BLAS_nthreads') > 0 else cpu_count() if self.get_config('BLAS_nthreads') == -1 else ERROR('Number of BLAS threads must be positive or -1')
 
         self.set_config('fit_time', None)
         totVoxels = np.count_nonzero(self.niiMASK_img)
-        LOG( f'\n-> Fitting "{self.model.name}" model to {totVoxels} voxels (using {n_jobs} job{"s" if n_jobs>1 else ""}):' )
+        # LOG( f'\n-> Fitting "{self.model.name}" model to {totVoxels} voxels (using {self.nthreads} thread{"s" if self.nthreads > 1 else ""}):' )
 
         # setup fitting directions
         peaks_filename = self.get_config('peaks_filename')
         if peaks_filename is None :
-            DIRs = np.zeros( [self.get_config('dim')[0], self.get_config('dim')[1], self.get_config('dim')[2], 3], dtype=np.float32 )
-            nDIR = 1
             if self.get_config('doMergeB0'):
                 gtab = gradient_table( np.hstack((0,self.scheme.b[self.scheme.dwi_idx])), np.vstack((np.zeros((1,3)),self.scheme.raw[self.scheme.dwi_idx,:3])) )
             else:
                 gtab = gradient_table( self.scheme.b, self.scheme.raw[:,:3] )
-            DTI = dti.TensorModel( gtab )
+            DTI = dti.TensorModel( gtab, fit_method=self.get_config('DTI_fit_method'))
         else :
             if not isfile( pjoin(self.get_config('DATA_path'), peaks_filename) ):
                 ERROR( 'PEAKS file not found' )
             niiPEAKS = nibabel.load( pjoin( self.get_config('DATA_path'), peaks_filename) )
-            DIRs = niiPEAKS.get_fdata().astype(np.float32)
-            nDIR = np.floor( DIRs.shape[3]/3 )
-            PRINT('\t* peaks dim = %d x %d x %d x %d' % DIRs.shape[:4])
-            if DIRs.shape[:3] != self.niiMASK_img.shape[:3] :
+            self.DIRs = niiPEAKS.get_fdata().astype(np.float32)
+            PRINT('\t* peaks dim = %d x %d x %d x %d' % self.DIRs.shape[:4])
+            if self.DIRs.shape[:3] != self.niiMASK_img.shape[:3]:
                 ERROR( 'PEAKS geometry does not match with DWI data' )
             DTI = None
-
-        # setup other output files
-        MAPs = np.zeros( [self.get_config('dim')[0], self.get_config('dim')[1], self.get_config('dim')[2], len(self.model.maps_name)], dtype=np.float32 )
 
         # fit the model to the data
         # =========================
         t = time.time()
+        # NOTE binary mask indexing
+        self.y = self.niiDWI_img[self.niiMASK_img==1, :].astype(np.double)
+        self.y[self.y < 0] = 0
 
-        ix, iy, iz = np.nonzero(self.niiMASK_img)
-        n_per_thread = np.floor(totVoxels / n_jobs)
-        idx = np.arange(0, totVoxels+1, n_per_thread, dtype=np.int32)
-        idx[-1] = totVoxels
+        # precompute directions
+        LOG(f"\n-> Estimating principal directions ({self.get_config('DTI_fit_method')}):")
+        if not self.get_config('doDirectionalAverage') and DTI is not None:
+            with ProgressBar(disable=get_verbose()<3):
+                self.DIRs = np.squeeze(DTI.fit(self.y).directions)
+        self.set_config('dirs_precomputing_time', time.time()-t)
+        LOG( '   [ %s ]' % ( time.strftime("%Hh %Mm %Ss", time.gmtime(self.get_config('dirs_precomputing_time')) ) ) )
 
-        estimates = Parallel(n_jobs=n_jobs)(
-            fit_voxel(self, ix[i], iy[i], iz[i], DIRs[ix[i],iy[i],iz[i],:], DTI)
-            for i in tqdm(range(totVoxels), ncols=70, bar_format='   |{bar}| {percentage:4.1f}%', disable=(get_verbose()<3))
-        )
-
+        t = time.time()
+        LOG(f"\n-> Fitting '{self.model.name}' model to {totVoxels} voxels (using {self.nthreads} thread{'s' if self.nthreads > 1 else ''}):")
+        # call the fit() method of the actual model
+        with self._controller.limit(limits=self.BLAS_nthreads, user_api='blas'):
+            results = self.model.fit(self)
         self.set_config('fit_time', time.time()-t)
         LOG( '   [ %s ]' % ( time.strftime("%Hh %Mm %Ss", time.gmtime(self.get_config('fit_time')) ) ) )
+        # =========================
 
         # store results
         self.RESULTS = {}
-
-        for i in range(totVoxels) :
-            MAPs[ix[i],iy[i],iz[i],:] = estimates[i][0]
-            DIRs[ix[i],iy[i],iz[i],:] = estimates[i][1]
-        self.RESULTS['DIRs']  = DIRs
-        self.RESULTS['MAPs']  = MAPs
-
+        # estimates (maps)
+        self.RESULTS['MAPs'] = np.zeros([self.get_config('dim')[0], self.get_config('dim')[1], self.get_config('dim')[2], len(self.model.maps_name)], dtype=np.float32)
+        self.RESULTS['MAPs'][self.niiMASK_img==1, :] = results['estimates']
+        # directions
+        self.RESULTS['DIRs'] = np.zeros([self.get_config('dim')[0], self.get_config('dim')[1], self.get_config('dim')[2], 3], dtype=np.float32)
+        self.RESULTS['DIRs'][self.niiMASK_img==1, :] = self.DIRs
+        # fitting error
+        if self.get_config('doComputeRMSE') :
+            self.RESULTS['RMSE'] = np.zeros([self.get_config('dim')[0], self.get_config('dim')[1], self.get_config('dim')[2]], dtype=np.float32)
+            self.RESULTS['RMSE'][self.niiMASK_img==1] = results['rmse']
         if self.get_config('doComputeNRMSE') :
-            NRMSE = np.zeros( [self.get_config('dim')[0], self.get_config('dim')[1], self.get_config('dim')[2]], dtype=np.float32 )
-            for i in range(totVoxels) :
-                NRMSE[ix[i],iy[i],iz[i]] = estimates[i][2]
-            self.RESULTS['NRMSE'] = NRMSE
-
-        if self.get_config('doSaveCorrectedDWI') :
-            DWI_corrected = np.zeros(self.niiDWI.shape, dtype=np.float32)
-            for i in range(totVoxels):
-                DWI_corrected[ix[i],iy[i],iz[i],:] = estimates[i][3]
-            self.RESULTS['DWI_corrected'] = DWI_corrected
+            self.RESULTS['NRMSE'] = np.zeros([self.get_config('dim')[0], self.get_config('dim')[1], self.get_config('dim')[2]], dtype=np.float32)
+            self.RESULTS['NRMSE'][self.niiMASK_img==1] = results['nrmse']
+        # Modulated NDI and ODI maps (NODDI)
+        if self.model.name == 'NODDI' and self.get_config('doSaveModulatedMaps'):
+            self.RESULTS['MAPs_mod'] = np.zeros([self.get_config('dim')[0], self.get_config('dim')[1], self.get_config('dim')[2], 2], dtype=np.float32)
+            self.RESULTS['MAPs_mod'][self.niiMASK_img==1, :] = results['estimates_mod']
+        # corrected DWI (Free-Water)
+        if self.model.name == 'Free-Water' and self.get_config('doSaveCorrectedDWI') :
+            y_corrected = results['y_corrected']
+            if self.get_config('doNormalizeSignal') and self.scheme.b0_count > 0:
+                y_corrected = y_corrected * np.reshape(self.mean_b0s[self.niiMASK_img==1], (-1, 1))
+            if self.get_config('doKeepb0Intact') and self.scheme.b0_count > 0:
+                y_corrected[:, self.scheme.b0_idx] = self.y[:, self.scheme.b0_idx] * np.reshape(self.mean_b0s[self.niiMASK_img==1], (-1, 1))
+            self.RESULTS['DWI_corrected'] = np.zeros(self.niiDWI.shape, dtype=np.float32)
+            self.RESULTS['DWI_corrected'][self.niiMASK_img==1, :] = y_corrected
 
 
     def save_results( self, path_suffix = None, save_dir_avg = False ) :
@@ -547,24 +519,23 @@ class Evaluation :
             if path_suffix :
                 RESULTS_path = RESULTS_path +'_'+ path_suffix
             self.RESULTS['RESULTS_path'] = RESULTS_path
-            LOG( f'\n-> Saving output to "{RESULTS_path}/*":' )
-
-            # delete previous output
+            LOG( f'\n-> Saving output to "{pjoin(RESULTS_path, "*")}":' )
             RESULTS_path = pjoin( self.get_config('DATA_path'), RESULTS_path )
         else:
             RESULTS_path = self.get_config('OUTPUT_path')
             if path_suffix :
                 RESULTS_path = RESULTS_path +'_'+ path_suffix
             self.RESULTS['RESULTS_path'] = RESULTS_path
-            LOG( f'\n-> Saving output to "{RESULTS_path}/*":' )
+            LOG( f'\n-> Saving output to "{pjoin(RESULTS_path, "*")}":' )
 
+        # delete previous output
         if not exists( RESULTS_path ) :
             makedirs( RESULTS_path )
         else :
             for f in glob.glob( pjoin(RESULTS_path,'*') ) :
                 remove( f )
 
-        # configuration
+        # configuration file
         PRINT('\t- configuration', end=' ')
         with open( pjoin(RESULTS_path,'config.pickle'), 'wb+' ) as fid :
             pickle.dump( self.CONFIG, fid, protocol=2 )
@@ -577,7 +548,7 @@ class Evaluation :
 
         # estimated orientations
         if not self.get_config('doDirectionalAverage'):
-            PRINT('\t- FIT_dir.nii.gz', end=' ')
+            PRINT('\t- fit_dir.nii.gz', end=' ')
             niiMAP_img = self.RESULTS['DIRs']
             niiMAP     = nibabel.Nifti1Image( niiMAP_img, affine, hdr )
             niiMAP_hdr = niiMAP.header if nibabel.__version__ >= '2.0.0' else niiMAP.get_header()
@@ -585,12 +556,23 @@ class Evaluation :
             niiMAP_hdr['cal_max'] = 1
             niiMAP_hdr['scl_slope'] = 1
             niiMAP_hdr['scl_inter'] = 0
-            nibabel.save( niiMAP, pjoin(RESULTS_path, 'FIT_dir.nii.gz') )
+            nibabel.save( niiMAP, pjoin(RESULTS_path, 'fit_dir.nii.gz') )
             PRINT(' [OK]')
 
         # fitting error
+        if self.get_config('doComputeRMSE') :
+            PRINT('\t- fit_RMSE.nii.gz', end=' ')
+            niiMAP_img = self.RESULTS['RMSE']
+            niiMAP     = nibabel.Nifti1Image( niiMAP_img, affine, hdr )
+            niiMAP_hdr = niiMAP.header if nibabel.__version__ >= '2.0.0' else niiMAP.get_header()
+            niiMAP_hdr['cal_min'] = 0
+            niiMAP_hdr['cal_max'] = 1
+            niiMAP_hdr['scl_slope'] = 1
+            niiMAP_hdr['scl_inter'] = 0
+            nibabel.save( niiMAP, pjoin(RESULTS_path, 'fit_RMSE.nii.gz') )
+            PRINT(' [OK]')
         if self.get_config('doComputeNRMSE') :
-            PRINT('\t- FIT_nrmse.nii.gz', end=' ')
+            PRINT('\t- fit_NRMSE.nii.gz', end=' ')
             niiMAP_img = self.RESULTS['NRMSE']
             niiMAP     = nibabel.Nifti1Image( niiMAP_img, affine, hdr )
             niiMAP_hdr = niiMAP.header if nibabel.__version__ >= '2.0.0' else niiMAP.get_header()
@@ -598,25 +580,26 @@ class Evaluation :
             niiMAP_hdr['cal_max'] = 1
             niiMAP_hdr['scl_slope'] = 1
             niiMAP_hdr['scl_inter'] = 0
-            nibabel.save( niiMAP, pjoin(RESULTS_path, 'FIT_nrmse.nii.gz') )
+            nibabel.save( niiMAP, pjoin(RESULTS_path, 'fit_NRMSE.nii.gz') )
             PRINT(' [OK]')
 
+        # corrected DWI (Free-Water)
         if self.get_config('doSaveCorrectedDWI') :
             if self.model.name == 'Free-Water' :
-                PRINT('\t- dwi_fw_corrected.nii.gz', end=' ')
+                PRINT('\t- DWI_corrected.nii.gz', end=' ')
                 niiMAP_img = self.RESULTS['DWI_corrected']
                 niiMAP     = nibabel.Nifti1Image( niiMAP_img, affine, hdr )
                 niiMAP_hdr = niiMAP.header if nibabel.__version__ >= '2.0.0' else niiMAP.get_header()
                 niiMAP_hdr['cal_min'] = 0
                 niiMAP_hdr['cal_max'] = 1
-                nibabel.save( niiMAP, pjoin(RESULTS_path, 'dwi_fw_corrected.nii.gz') )
+                nibabel.save( niiMAP, pjoin(RESULTS_path, 'DWI_corrected.nii.gz') )
                 PRINT(' [OK]')
             else :
                 WARNING( f'"doSaveCorrectedDWI" option not supported for "{self.model.name}" model' )
 
         # voxelwise maps
         for i in range( len(self.model.maps_name) ) :
-            PRINT(f'\t- FIT_{self.model.maps_name[i]}.nii.gz', end=' ')
+            PRINT(f'\t- fit_{self.model.maps_name[i]}.nii.gz', end=' ')
             niiMAP_img = self.RESULTS['MAPs'][:,:,:,i]
             niiMAP     = nibabel.Nifti1Image( niiMAP_img, affine, hdr )
             niiMAP_hdr = niiMAP.header if nibabel.__version__ >= '2.0.0' else niiMAP.get_header()
@@ -625,8 +608,28 @@ class Evaluation :
             niiMAP_hdr['cal_max'] = niiMAP_img.max()
             niiMAP_hdr['scl_slope'] = 1
             niiMAP_hdr['scl_inter'] = 0
-            nibabel.save( niiMAP, pjoin(RESULTS_path, f'FIT_{self.model.maps_name[i]}.nii.gz' ) )
+            nibabel.save( niiMAP, pjoin(RESULTS_path, f'fit_{self.model.maps_name[i]}.nii.gz' ) )
             PRINT(' [OK]')
+
+        # modulated NDI and ODI maps (NODDI)
+        if self.get_config('doSaveModulatedMaps'):
+            if self.model.name == 'NODDI':
+                mod_maps = [name + '_modulated' for name in self.model.maps_name[:2]]
+                descr = [descr + ' modulated' for descr in self.model.maps_descr[:2]]
+                for i in range(len(mod_maps)):
+                    PRINT(f'\t- fit_{mod_maps[i]}.nii.gz', end=' ')
+                    niiMAP_img = self.RESULTS['MAPs_mod'][:,:,:,i]
+                    niiMAP     = nibabel.Nifti1Image( niiMAP_img, affine, hdr )
+                    niiMAP_hdr = niiMAP.header if nibabel.__version__ >= '2.0.0' else niiMAP.get_header()
+                    niiMAP_hdr['descrip'] = descr[i] + f' (AMICO v{self.get_config("version")})'
+                    niiMAP_hdr['cal_min'] = niiMAP_img.min()
+                    niiMAP_hdr['cal_max'] = niiMAP_img.max()
+                    niiMAP_hdr['scl_slope'] = 1
+                    niiMAP_hdr['scl_inter'] = 0
+                    nibabel.save( niiMAP, pjoin(RESULTS_path, f'fit_{mod_maps[i]}.nii.gz' ) )
+                    PRINT(' [OK]')
+            else:
+                WARNING(f'"doSaveModulatedMaps" option not supported for "{self.model.name}" model')
 
         # Directional average signal
         if save_dir_avg:
@@ -639,7 +642,7 @@ class Evaluation :
                 PRINT(' [OK]')
 
                 PRINT('\t- dir_avg.scheme', end=' ')
-                np.savetxt( pjoin(RESULTS_path, 'dir_avg.scheme' ), self.scheme.get_table(), fmt="%.06f", delimiter="\t", header=f"VERSION: {self.scheme.version}", comments='' )
+                np.savetxt( pjoin(RESULTS_path, 'dir_avg.scheme' ), self.scheme.get_table(), fmt="%.06f", delimiter="\t", header=f'VERSION: {self.scheme.version}', comments='' )
                 PRINT(' [OK]')
             else:
                 WARNING('The directional average signal was not created (The option doDirectionalAverage is False).')