+++ amico/synthesis.py
@@ -1,7 +1,6 @@
 from abc import ABC, abstractmethod
 import numpy as np
-import numpy.matlib as matlib
-import scipy
+from scipy.special import erf, erfi, lpmv
 from amico.util import ERROR
 
 # Limits the required precision in gpd sum
@@ -26,7 +25,7 @@ def _gpd_sum(am, big_delta, small_delta, diff, radius, n):
         sum += term
         if term < _REQUIRED_PRECISION * sum:
             break
-    return big_delta, small_delta, diff, radius, sum
+    return sum
 
 def _scheme2noddi(scheme):
     protocol = {}
@@ -258,7 +257,7 @@ class Ball(BaseTensor):
         return super()._get_signal(evals)
 
 # SPHERE
-class SphereGPD():
+class SphereGPD:
     """
     Class to generate the signal from a sphere compartment with GPD approximation.
 
@@ -337,12 +336,16 @@ class SphereGPD():
                 g_mods = g_dir * g
                 g_mod = np.sqrt(np.dot(g_mods, g_mods))
                 if big_delta != self._last_big_delta or small_delta != self._last_small_delta or diff != self._last_diff or radius != self._last_radius:
-                    self._last_big_delta, self._last_small_delta, self._last_diff, self._last_radius, self._last_sum = _gpd_sum(am, big_delta, small_delta, diff, radius, 2)
+                    self._last_big_delta = big_delta
+                    self._last_small_delta = small_delta
+                    self._last_diff = diff
+                    self._last_radius = radius
+                    self._last_sum = _gpd_sum(am, big_delta, small_delta, diff, radius, 2)
                 signal[i] = np.exp(-2 * _GAMMA * _GAMMA * g_mod * g_mod * self._last_sum)
         return signal
 
 # ASTROSTICKS
-class Astrosticks():
+class Astrosticks:
     """
     Class to generate the signal from an astrosticks compartment (sticks with distributed orientations).
 
@@ -386,11 +389,11 @@ class Astrosticks():
             else:
                 l_perp = 0
                 l_par = -b / (g * g) * diff
-                signal[i] = np.sqrt(np.pi) * 1 / (2 * g * np.sqrt(l_perp - l_par)) * np.exp(g * g * l_perp) * scipy.special.erf(g * np.sqrt(l_perp - l_par))
+                signal[i] = np.sqrt(np.pi) * 1 / (2 * g * np.sqrt(l_perp - l_par)) * np.exp(g * g * l_perp) * erf(g * np.sqrt(l_perp - l_par))
         return signal
 
 # CYLINDER
-class CylinderGPD():
+class CylinderGPD:
     """
     Class to generate the signal from a cylinder compartment with GPD approximation.
 
@@ -477,7 +480,11 @@ class CylinderGPD():
                     unit_gn = gn / (g_mod * n_mod)
                 omega = np.arccos(unit_gn)
                 if big_delta != self._last_big_delta or small_delta != self._last_small_delta or diff != self._last_diff or radius != self._last_radius:
-                    self._last_big_delta, self._last_small_delta, self._last_diff, self._last_radius, self._last_sum = _gpd_sum(am, big_delta, small_delta, diff, radius, 1)
+                    self._last_big_delta = big_delta
+                    self._last_small_delta = small_delta
+                    self._last_diff = diff
+                    self._last_radius = radius
+                    self._last_sum = _gpd_sum(am, big_delta, small_delta, diff, radius, 1)
                 sr_perp = np.exp(-2 * _GAMMA * _GAMMA * g_mod * g_mod * np.sin(omega) * np.sin(omega) * self._last_sum)
                 t = big_delta - small_delta / 3
                 sr_par = np.exp(-t * (_GAMMA * small_delta * g_mod * np.cos(omega) * (_GAMMA * small_delta * g_mod * np.cos(omega))) * diff)
@@ -485,7 +492,7 @@ class CylinderGPD():
         return signal
 
 # NODDI
-class NODDIIntraCellular():
+class NODDIIntraCellular:
     def __init__(self, scheme):
         self.scheme = scheme
         self.protocol_hr = _scheme2noddi(self.scheme)
@@ -522,7 +529,7 @@ class NODDIIntraCellular():
 
         # Compute the spherical harmonic coefficients of the Watson's distribution
         coeff = self._watson_SH_coeff(kappa)
-        coeffMatrix = matlib.repmat(coeff, l_q, 1)
+        coeffMatrix = np.tile(coeff, (l_q, 1))
 
         # Compute the dot product between the symmetry axis of the Watson's distribution
         # and the gradient direction
@@ -541,7 +548,7 @@ class NODDIIntraCellular():
 
         # Compute the SH values at cosTheta
         sh = np.zeros(coeff.shape)
-        shMatrix = matlib.repmat(sh, l_q, 1)
+        shMatrix = np.tile(sh, (l_q, 1))
         for i in range(7):
             shMatrix[:,i] = np.sqrt((i+1 - .75)/np.pi)
             # legendre function returns coefficients of all m from 0 to l
@@ -550,7 +557,7 @@ class NODDIIntraCellular():
             # cosTheta is expected to be a COLUMN vector.
             tmp = np.zeros((l_q))
             for pol_i in range(l_q):
-                tmp[pol_i] = scipy.special.lpmv(0, 2*i, cosTheta[pol_i])
+                tmp[pol_i] = lpmv(0, 2*i, cosTheta[pol_i])
             shMatrix[:,i] = shMatrix[:,i]*tmp
 
         E = np.sum(lgi*coeffMatrix*shMatrix, 1)
@@ -600,7 +607,7 @@ class NODDIIntraCellular():
 
         I = np.zeros((len(Lpmp),mn))
         sqrtx = np.sqrt(Lpmp[exact])
-        I[exact,0] = np.sqrt(np.pi)*scipy.special.erf(sqrtx)/sqrtx
+        I[exact,0] = np.sqrt(np.pi)*erf(sqrtx)/sqrtx
         dx = 1.0/Lpmp[exact]
         emx = -np.exp(-Lpmp[exact])
         for i in range(1,mn):
@@ -674,7 +681,7 @@ class NODDIIntraCellular():
         k6 = k5*kappa
         k7 = k6*kappa
 
-        erfik = scipy.special.erfi(sk)
+        erfik = erfi(sk)
         ierfik = 1/erfik
         ek = np.exp(kappa)
         dawsonk = 0.5*np.sqrt(np.pi)*erfik/ek
@@ -751,7 +758,7 @@ class NODDIIntraCellular():
             C[6] = 128*np.sqrt(np.pi)*k6/152108775
         return C
 
-class NODDIExtraCellular():
+class NODDIExtraCellular:
     def __init__(self, scheme):
         self.scheme = scheme
         self.protocol_hr = _scheme2noddi(self.scheme)
@@ -792,7 +799,7 @@ class NODDIExtraCellular():
             dw[1] = dParP2dPerp/3.0 - 2.0*dParMdPerp*kappa/45.0 - 4.0*dParMdPerp*k2/945.0
         else:
             sk = np.sqrt(kappa)
-            dawsonf = 0.5*np.exp(-kappa)*np.sqrt(np.pi)*scipy.special.erfi(sk)
+            dawsonf = 0.5*np.exp(-kappa)*np.sqrt(np.pi)*erfi(sk)
             factor = sk/dawsonf
             dw[0] = (-dParMdPerp+2.0*dPerp*kappa+dParMdPerp*factor)/(2.0*kappa)
             dw[1] = (dParMdPerp+2.0*(dPar+dPerp)*kappa-dParMdPerp*factor)/(4.0*kappa)
@@ -818,7 +825,7 @@ class NODDIExtraCellular():
         E=np.exp(-bval*((dPar - dPerp)*cosThetaSq + dPerp))
         return E
 
-class NODDIIsotropic():
+class NODDIIsotropic:
     def __init__(self, scheme):
         self.scheme = scheme
         self.protocol_hr = _scheme2noddi(self.scheme)